Move out validateArchive to its own function
So that I can use it in one more spot.
