+++ test/validateArchive-test.js
